drivers: ethernet: stm32: move into own dir
move stm32 ethernet drivers into own directory.
