+++ drivers/ethernet/Kconfig
@@ -72,7 +72,6 @@ source "drivers/ethernet/Kconfig.sam_gmac"
 source "drivers/ethernet/Kconfig.smsc911x"
 source "drivers/ethernet/Kconfig.smsc91x"
 source "drivers/ethernet/Kconfig.stellaris"
-source "drivers/ethernet/Kconfig.stm32_hal"
 source "drivers/ethernet/Kconfig.sy1xx_mac"
 source "drivers/ethernet/Kconfig.test"
 source "drivers/ethernet/Kconfig.virtio_net"
@@ -91,6 +90,7 @@ source "drivers/ethernet/intel/Kconfig"
 source "drivers/ethernet/nxp/Kconfig"
 source "drivers/ethernet/offload/Kconfig"
 source "drivers/ethernet/phy/Kconfig"
+source "drivers/ethernet/stm32/Kconfig"
 # zephyr-keep-sorted-stop
 
 endif # "Ethernet Drivers"
