+++ drivers/ethernet/stm32/Kconfig
@@ -2,6 +2,7 @@
 
 # Copyright (c) 2017 Erwin Rol <erwin@erwinrol.com>
 # Copyright (c) 2020 Alexander Kozhinov <ak.alexander.kozhinov@gmail.com>
+# SPDX-FileCopyrightText: Copyright The Zephyr Project Contributors
 # SPDX-License-Identifier: Apache-2.0
 
 menuconfig ETH_STM32_HAL
