+++ drivers/ethernet/stm32/eth_stm32_hal_priv.h
