+++ drivers/ethernet/stm32/eth_stm32_hal_v1.c
@@ -20,7 +20,7 @@
 #include <stdbool.h>
 #include <errno.h>
 
-#include "eth.h"
+#include "../eth.h"
 #include "eth_stm32_hal_priv.h"
 
 #define DT_DRV_COMPAT st_stm32_ethernet
